+++ test/libsolidity/astPropertyTests/transient_data_location.sol
@@ -1,0 +1,14 @@
+contract C {
+    /// TransientDataLocation: storageLocation
+    /// TransientVarName: name
+    uint transient x;
+    /// StorageDataLocation: storageLocation
+    /// StorageVarName: name
+    uint y;
+
+}
+// ----
+// TransientDataLocation: transient
+// TransientVarName: x
+// StorageDataLocation: default
+// StorageVarName: y